Changed URL for user data - Now under '/game/data/user'
