+++ mysite/api/serializers.py
@@ -1,7 +1,5 @@
 from rest_framework import serializers
-from django.contrib.auth.models import User
 from .models import UserMessage
-from accounts.models import UserInfo
 
 class MessageSerializer(serializers.ModelSerializer):
     sender_username = serializers.SerializerMethodField()
@@ -14,11 +12,3 @@ class MessageSerializer(serializers.ModelSerializer):
         fields = ['id', 'sender', 'receiver', 'content', 'timestamp', 'sender_username']
 
 
-class UserInfoSerializer(serializers.ModelSerializer):
-    user_username = serializers.SerializerMethodField()
-
-    def get_user_username(self, obj):
-        return obj.user.username
-    class Meta:
-        model = UserInfo
-        fields = ['id', 'user_username', 'coins', 'highscore', 'cumulativeScore']
