+++ mysite/api/urls.py
@@ -6,5 +6,4 @@ app_name = 'api'
 
 urlpatterns = [
     path('messages/<str:username>/', MessageData.as_view(), name='message_data'),
-    path('data/user/', UserData.as_view(), name='user_data'),
 ]
