+++ mysite/api/views.py
@@ -5,8 +5,7 @@ from rest_framework.response import Response
 
 # Non-static imports
 from .models import UserMessage
-from .serializers import MessageSerializer, UserInfoSerializer
-from accounts.models import UserInfo
+from .serializers import MessageSerializer
 
 """
  * A REST API view to handle messages between users
@@ -40,26 +39,3 @@ class MessageData(generics.GenericAPIView):
 
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class UserData(generics.GenericAPIView):
-    serializer_class = UserInfoSerializer
-    def get(self, request, *args, **kwargs):
-        user_object = User.objects.filter(username__iexact=self.request.user).first()
-        if user_object:
-            queryset = UserInfo.objects.get(user=user_object)
-            serializer = UserInfoSerializer(queryset, many=False)
-
-            return Response(serializer.data, status=status.HTTP_200_OK)
-        return Response({}, status=status.HTTP_404_NOT_FOUND)
-    def post(self, request, *args, **kwargs):
-        user_object = User.objects.filter(username__iexact=self.request.user).first()
-        if not user_object:
-            return Response({"error": "User not found"}, status=status.HTTP_404_NOT_FOUND)
-
-        user_info_object = UserInfo.objects.get(user=user_object)
-        serializer = self.get_serializer(user_info_object, data=request.data, partial=True)
-
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
-
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
