+++ mysite/game/urls.py
@@ -5,4 +5,5 @@ from .views import *
 urlpatterns = [
     path('', game_view, name='game'),
     path('<str:filename>/', GameFileView.as_view(), name='game__get_file'),
+    path('data/user/', UserData.as_view(), name='user_data'),
 ]
